+++ majordome/simulation/__init__.py
@@ -1,7 +1,8 @@
 # -*- coding: utf-8 -*-
-import importlib
 
-_LAZY_EXPORTS = {
+from .._imports import setup_lazy_exports
+
+__getattr__, __dir__ = setup_lazy_exports(__name__, globals(), {
     # elmer:
     "ConstantTimeStepInterval": ".elmer",
     "TimeStepAccumulator": ".elmer",
@@ -31,21 +32,4 @@ _LAZY_EXPORTS = {
     "hexagon_points_xy": ".meshing",
     "square_points_xy": ".meshing",
     "get_extrusion_tags": ".meshing",
-}
-
-__all__ = list(_LAZY_EXPORTS.keys())
-
-
-def __getattr__(name: str):
-    if name in _LAZY_EXPORTS:
-        submodule_path = _LAZY_EXPORTS[name]
-        submodule = importlib.import_module(submodule_path, __package__)
-        exported_item = getattr(submodule, name)
-        globals()[name] = exported_item
-        return exported_item
-
-    raise AttributeError(f"module '{__name__}' has no attribute '{name}'")
-
-
-def __dir__():
-    return list(globals().keys()) + __all__
+})